requirements.txt
google-cloud-firestore==2.13.1
flask==3.0.0
functions-framework==3.5.0